Add 0.2.01 firmware
Revision 0.2.01 corrects a bug in the register read logic. Internally, the
modules within the fabric output data only when they are addressed, and each
module's output is OR'd together to create one internal read bus. In one
of the modules, there was a bug that caused it to output data when another
module was being addressed. This bug is fixed in revision 0.2.01

Tested:
  AXON-IMX6: TBD
  AXON-IMX8MM: Tested on Axon-Wizard with AXON-IMX8MM loaded with F01 fabric (2100 LUTs)